Moving test around
